refactor!: Move `fs/utils` package to `fsutils` package namespace
