+++ fsutils/fsutils.go
@@ -2,7 +2,7 @@
 // Copyright (c) 2025, Unikraft GmbH and The KraftKit Authors.
 // Licensed under the BSD-3-Clause License (the "License").
 // You may not use this file except in compliance with the License.
-package utils
+package fsutils
 
 import (
 	"archive/tar"
